# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.0.120:8000",
    "paths": {
...
        "/api/v1/kho": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter ware house based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "ware house"
                ],
                "summary": "Filter Ware House",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing ware house entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ware house"
                ],
                "summary": "Update WareHouse",
                "parameters": [
                    {
                        "description": "Updated ware house data",
                        "name": "Ware_House",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Kho_update"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new ware house entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ware house"
                ],
                "summary": "Create Ware House",
                "parameters": [
                    {
                        "description": "ware house data",
                        "name": "Ware_House",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Kho_create"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/kho/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing ware house entry",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ware house"
                ],
                "summary": "Delete Ware House",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ware House ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/loai-san-pham": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter product type based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "product type"
                ],
                "summary": "Filter Product Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing product type entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product type"
                ],
                "summary": "Update Product Type",
                "parameters": [
                    {
                        "description": "Product Type Update Data",
                        "name": "UpdateProductType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Loai_san_pham_update"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: cap nhat loai san pham thanh cong",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "message: error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product type entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product type"
                ],
                "summary": "Create Product Type",
                "parameters": [
                    {
                        "description": "Product Type Create Data",
                        "name": "CreateProductType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Loai_san_pham_create"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: Loai_san_pham_create, message: them loai san pham thanh cong",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "message: error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/loai-san-pham/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing product type entry",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product type"
                ],
                "summary": "Delete Product Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product type ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "requests.Kho_create": {
            "type": "object",
            "required": [
                "dia_chi",
                "ten"
            ],
            "properties": {
                "dia_chi": {
                    "type": "string"
                },
                "ten": {
                    "type": "string"
                }
            }
        },
        "requests.Kho_update": {
            "type": "object",
            "required": [
                "dia_chi",
                "id",
                "ten"
            ],
            "properties": {
                "dia_chi": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ten": {
                    "type": "string"
                }
            }
        },
...
        "requests.Loai_san_pham_create": {
            "type": "object",
            "required": [
                "hinh_anh",
                "ten"
            ],
            "properties": {
                "hinh_anh": {
                    "type": "string"
                },
                "ten": {
                    "type": "string"
                }
            }
        },
        "requests.Loai_san_pham_update": {
            "type": "object",
            "required": [
                "id",
                "ten"
            ],
            "properties": {
                "hinh_anh": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ten": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}